requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file requirements.txt
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
absl-py=2.0.0=pypi_0
accelerate=0.24.0=pypi_0
allennlp=2.10.1=pypi_0
anykeystore=0.2=pypi_0
apex=0.9.10.dev0=pypi_0
asttokens=2.4.0=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
base58=2.1.1=pypi_0
bitsandbytes=0.41.1=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
bleurt=0.0.2=pypi_0
boto3=1.34.113=pypi_0
botocore=1.34.113=pypi_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.7.22=hbcca054_0
cached-path=1.1.6=pypi_0
cachetools=5.3.2=pypi_0
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.4=pyhd8ed1ab_0
commonmark=0.9.1=pypi_0
cryptacular=1.6.2=pypi_0
cryptography=41.0.3=py310hdda0065_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
datasets=2.18.0=pypi_0
debugpy=1.8.0=py310hc6cd4ac_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
docker-pycreds=0.4.0=pypi_0
einops=0.7.0=pypi_0
en-core-web-sm=3.5.0=pypi_0
evaluate=0.4.2=pypi_0
exceptiongroup=1.1.3=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
fairscale=0.4.6=pypi_0
fastcoref=2.1.6=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.7.1=pypi_0
flatbuffers=23.5.26=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2023.10.0=pypi_0
gast=0.5.4=pypi_0
giflib=5.2.1=h5eee18b_3
gitignore-parser=0.1.11=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py310heeb90bb_0
gnutls=3.6.15=he1e5248_0
google-ai-generativelanguage=0.6.4=pypi_0
google-api-python-client=2.130.0=pypi_0
google-auth=2.24.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-auth-oauthlib=1.1.0=pypi_0
google-generativeai=0.5.4=pypi_0
google-pasta=0.2.0=pypi_0
gpustat=1.1.1=pypi_0
gputil=1.4.0=pypi_0
greenlet=3.0.3=pypi_0
groq=0.8.0=pypi_0
grpcio=1.59.3=pypi_0
h5py=3.10.0=pypi_0
httplib2=0.22.0=pypi_0
huggingface-hub=0.17.1=pypi_0
hupper=1.12.1=pypi_0
idna=3.4=py310h06a4308_0
imgkit=1.2.3=pypi_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
iniconfig=2.0.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
ipykernel=6.25.2=pyh2140261_0
ipython=8.15.0=pyh0d859eb_0
jedi=0.19.0=pyhd8ed1ab_0
jinja2=3.1.2=py310h06a4308_0
jiwer=3.0.4=pypi_0
jmespath=1.0.1=pypi_0
jpeg=9e=h5eee18b_1
jsonnet=0.20.0=pypi_0
jupyter_client=8.3.1=pyhd8ed1ab_0
jupyter_core=5.3.1=py310hff52083_0
keras=2.15.0=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libabseil=20230802.0=cxx17_h59595ed_3
libclang=16.0.6=pypi_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.7.2.10=0
libcurand=10.3.3.141=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libffi=3.4.2=h7f98852_5
libgcc-ng=13.2.0=h807b86a_0
libgomp=13.2.0=h807b86a_0
libidn2=2.3.4=h5eee18b_0
libnpp=11.8.0.86=0
libnsl=2.0.0=h5eee18b_0
libnvjpeg=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libprotobuf=4.23.4=hf27288f_6
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.43.0=h2797004_0
libstdcxx-ng=13.2.0=h7e041cc_0
libtasn1=4.19.0=h5eee18b_0
libtiff=4.4.0=hecacb30_2
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libzlib=1.2.13=hd590300_5
lmdb=1.4.1=pypi_0
lxml=5.2.2=pypi_0
lz4-c=1.9.4=h6a678d5_0
mantis=0.0.1=pypi_0
markdown=3.5.1=pypi_0
markupsafe=2.1.1=py310h7f8727e_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mkl=2023.1.0=h213fc3f_46343
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
ml-dtypes=0.2.0=pypi_0
more-itertools=10.2.0=pypi_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py310h06a4308_0
multiprocess=0.70.16=pypi_0
ncurses=6.4=hcb278e6_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py310h06a4308_0
nltk=3.8.1=pypi_0
numpy=1.25.2=py310h5f9d8c6_0
numpy-base=1.25.2=py310hb5e798b_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-ml-py=12.555.43=pypi_0
nvidia-nccl-cu12=2.19.3=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
oauthlib=3.2.2=pypi_0
openai=1.30.5=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=3.1.2=hd590300_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pastedeploy=3.1.0=pypi_0
pathtools=0.1.2=pypi_0
pbkdf2=1.3=pypi_0
pbr=2.1.0=pypi_0
peft=0.5.0=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310h6a678d5_0
pip=23.2.1=py310h06a4308_0
plaster=1.1.2=pypi_0
plaster-pastedeploy=1.0.1=pypi_0
platformdirs=3.10.0=pyhd8ed1ab_0
pluggy=1.5.0=pypi_0
portalocker=2.8.2=pypi_0
promise=2.3=pypi_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
proto-plus=1.24.0.dev1=pypi_0
protobuf=3.20.0=pypi_0
psutil=5.9.5=py310h1fa729e_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=16.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.8.2=pypi_0
pygments=2.16.1=pyhd8ed1ab_0
pyopenssl=23.2.0=py310h06a4308_0
pyparsing=3.1.1=pypi_0
pyramid=2.0.2=pypi_0
pyramid-mailer=0.15.1=pypi_0
pysocks=1.7.1=py310h06a4308_0
pytest=8.2.1=pypi_0
python=3.10.12=hd12c33a_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python3-openid=3.2.0=pypi_0
python_abi=3.10=3_cp310
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.1=py310h5bbb5d0_0
rapidfuzz=3.9.3=pypi_0
readline=8.2=h8228510_1
repoze-sendmail=4.4.1=pypi_0
requests=2.31.0=py310h06a4308_0
requests-oauthlib=1.3.1=pypi_0
rich=12.6.0=pypi_0
rouge=1.0.1=pypi_0
rouge-score=0.1.2=pypi_0
rsa=4.9=pypi_0
s3transfer=0.10.1=pypi_0
sacrebleu=2.4.2=pypi_0
sacremoses=0.1.1=pypi_0
safetensors=0.3.3=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.3=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=2.3.1=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=pyhd8ed1ab_0
shortuuid=1.0.13=pypi_0
six=1.16.0=pyh6c4a22f_0
spacy=3.3.3=pypi_0
sqlalchemy=2.0.27=pypi_0
sqlite=3.43.0=h2c6b66d_0
stack_data=0.6.2=pyhd8ed1ab_0
sympy=1.11.1=py310h06a4308_0
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.15.1=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorboardx=2.6.2.2=pypi_0
tensorflow=2.15.0=pypi_0
tensorflow-estimator=2.15.0=pypi_0
tensorflow-io-gcs-filesystem=0.34.0=pypi_0
termcolor=1.1.0=pypi_0
tf-slim=1.1.0=pypi_0
thinc=8.0.17=pypi_0
tiktoken=0.6.0=pypi_0
timm=0.9.16=pypi_0
tk=8.6.12=h27826a3_0
token-count=0.2.1=pypi_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
torch=1.12.1=pypi_0
torchaudio=2.0.2=py310_cu118
torchvision=0.13.1=pypi_0
tornado=6.3.3=py310h2372a71_0
tqdm=4.66.1=pypi_0
traitlets=5.10.0=pyhd8ed1ab_0
transaction=4.0=pypi_0
transformers=4.41.1=pypi_0
transformers-stream-generator=0.0.4=pypi_0
translationstring=1.4=pypi_0
triton=2.2.0=pypi_0
typer=0.4.2=pypi_0
typing-extensions=4.10.0=pypi_0
tzdata=2023c=h71feb2d_0
uritemplate=4.1.1=pypi_0
urllib3=1.26.16=py310h06a4308_0
velruse=1.1.1=pypi_0
venusian=3.1.0=pypi_0
wandb=0.12.21=pypi_0
wasabi=0.10.1=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
webob=1.8.7=pypi_0
werkzeug=3.0.1=pypi_0
wget=3.2=pypi_0
wheel=0.41.2=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
wtforms=3.1.2=pypi_0
wtforms-recaptcha=0.3.2=pypi_0
wxconv=1.0.0.0=pypi_0
x264=1!157.20191217=h7b6447c_0
xformers=0.0.26.post1=pypi_0
xxhash=3.4.1=pypi_0
xz=5.2.6=h166bdaf_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zope-deprecation=5.0=pypi_0
zope-interface=6.2=pypi_0
zope-sqlalchemy=3.1=pypi_0
zstd=1.5.2=ha4553b6_0
